+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## [2.3.1] - 2026-02-27
+
+### Added
+- Added Chinese (zh_cn) translation (By @ChujikYahus).
+
+## [2.3.0] - 2026-02-26
+
+### Added
+
+- Add item transfer support of traders in spirit jars (By @antessial).
+
+## [2.2.1] - 2026-02-06
+
+### Fixed
+
+- Fix sacrificial altar loot table. It should drop as an item when broken.
+
+## [2.2.0] - 2025-08-02
+
+### Added
+
+- Added support for Crystallizer spirits for auto item transfer to nearby inventories.
+- Added Sacrificial Altar - New Block that can be used to sacrifice mobs from Mob jars. It consumes 5000 Source and must
+  be placed under the golden sacrificial bowl
+
+### Changed
+
+- Bumped minimum version of Occultism and Neoforge
+
 ## [2.1.0] - 2024-12-23
 
 ### Added